requirements.txt
# Core Libraries
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.2
scikit-learn>=1.0.2
tqdm>=4.62.0
Pillow>=9.0.0

# CatBoost
catboost>=1.2.7

# CLIP
clip@git+https://github.com/openai/CLIP.git

# Optional: For FITS file handling
astropy>=5.0.4

# Jupyter for notebooks
jupyter>=1.0.0